+++ src/index.js
@@ -8,45 +8,21 @@
 // Importing dependencies
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { Route, NavLink, BrowserRouter as Router, Switch } from 'react-router-dom'
-import './index.css';
 import App from './App';
-import Login from './login';
-import Cart from './cart';
-import Products from './products';
-import Notfound from './notFound';
-import Header from "./header";
-import Footer from "./footer";
+import './index.css'
+// import cartReducer from './components/reducers/cartReducer';
+// import { Provider } from 'react-redux';
+// import { createStore } from 'redux';
 import * as serviceWorker from './serviceWorker';
 import 'bootstrap/dist/css/bootstrap.min.css'; // Importing React-Bootstrap Style Library
 
-// Creating contact data to be passed to the footer component
-var contactData = {
-  phone: "1-800-0000",
-  emailAddresses:["info@bookstore.com", "order-support@bookstore.com"],
-  street: "800 Franklin Street",
-  city: "Cityville",
-  state: "AZ",
-  zip: "11111"
-}
+// const store = createStore(cartReducer);
 
-// Creating the routing to link between pages
-const routing = (
-  <Router>
-    <div>
-      <Header />
-      <Switch>
-        <Route exact path="/" component={App} />
-        <Route path="/products" component={Products} />
-        <Route path="/cart" component={Cart} />
-        <Route path="/login" component={Login} />
-        <Route component={Notfound} />
-      </Switch>
-      <Footer contactData={contactData}/>
-    </div>
-  </Router>
-)
-
-ReactDOM.render(routing, document.getElementById('root'));
+ReactDOM.render(
+  // <Provider store={store}>
+  <App />
+  // </Provider>
+  , document.getElementById('root')
+);
 
 serviceWorker.unregister();
